+++ tomcat/src/main/java/nextstep/jwp/controller/StaticFileController.java
@@ -3,6 +3,7 @@ package nextstep.jwp.controller;
 import java.io.File;
 import java.net.URL;
 import nextstep.jwp.exception.HttpMethodNotAllowedException;
+import org.apache.coyote.http11.controller.AbstractController;
 import org.apache.coyote.http11.request.HttpRequest;
 import org.apache.coyote.http11.response.HttpResponse;
 import org.apache.coyote.http11.response.HttpResponse.ResponseBuilder;
