+++ tomcat/src/main/java/org/apache/coyote/http11/controller/RequestMapping.java
@@ -1,6 +1,10 @@
-package nextstep.jwp.controller;
+package org.apache.coyote.http11.controller;
 
 import java.util.Map;
+import nextstep.jwp.controller.IndexController;
+import nextstep.jwp.controller.LoginController;
+import nextstep.jwp.controller.RegisterController;
+import nextstep.jwp.controller.StaticFileController;
 import org.apache.coyote.http11.request.HttpRequest;
 
 public class RequestMapping {
